backend/api_olap/requirements.txt
flask 
flask-mysql
mysql-connector-python
python-dotenv
loguru
clickhouse-connect
